+++ CHANGELOG.md
@@ -4,10 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.4] - 2021-04-15
+### Changes
+- Fix RPC thread performance issue
+- log event\status submission issues
+
 ## [0.1.3] - 2021-04-15
+### Changes
 - fix submit interval update issue
 
 ## [0.1.2] - 2021-04-15
+### Changes
 - fix event-action path
 
 ## [0.1.1] - 2021-04-12